+++ .eslintrc.js
@@ -3,11 +3,7 @@ const path = require("path")
 module.exports = {
 	extends: [
 		path.join(__dirname, "node.js"),
-		path.join(__dirname, "mocha.js"),
-		path.join(__dirname, "import.js")
-	],
-
-	rules: {
-		"import/default": "off"
-	}
+		path.join(__dirname, "import.js"),
+		path.join(__dirname, "jest.js")
+	]
 }