+++ jest.js
@@ -1,0 +1,11 @@
+module.exports = {
+	plugins: ["jest"],
+
+	rules: {
+
+		// Jest
+		"jest/no-disabled-tests": "error",
+		"jest/no-focused-tests": "error",
+		"jest/no-identical-title": "error"
+	}
+}